+++ source/utils/utils.h
@@ -495,7 +495,7 @@ namespace chord
 
 
 	// From https://github.com/TheCherno/Hazel/blob/master/Hazel/src/Hazel/Math/Math.cpp
-	extern bool decomposeTransform(const math::mat4& transform, math::vec3& translation, math::vec3& rotation, math::vec3& scale);
+	extern bool decomposeTransform(const math::dmat4& transform, math::dvec3& translation, math::dvec3& rotation, math::dvec3& scale);
 
 	static inline bool hasFlag(auto a, auto b)
 	{
